drm/i915: refactor & correct GEN6 PIPE_CONTROL code
The code in gen6_render_ring_flush() wasn't quite right; whereas the
workarounds were emitted with GFX_OP_PIPE_CONTROL(5) (allowing for a
QWord write), the final PIPE_CONTROL used GFX_OP_PIPE_CONTROL(4) even
though it would also have the PIPE_CONTROL_QW_WRITE flag set. This may
(should?) have resulted in the next instruction word being consumed as
data, or the PIPE_CONTROL not working as intended due to the flags-vs-
length mismatch.

The refactored code has a function which will emit a complete
PIPE_CONTROL sequence, with the correct number of extension DWords. The
emit-workarounds and emit-flush functions are then trivially expressed
as calls to the new low-level function.

XXX: to check: if this is used on GEN7+ => is GTT bit still in address?

Signed-off-by: Dave Gordon <david.s.gordon@intel.com>
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>