drm/i915: Introduce initial mode setting (v10)
For very specific use cases it may be desirable to have the kernel
driver do an initial mode set and/or splash screen type display without
having to enable a framebuffer console.

A primary goal for this feature is to have the display lit up as soon
as possible.  A non-goal is the ability to detect and adapt to the
display configurations. In other words, the  display configuration will
be a known/fixed configuration.  In addition, it is expected that the
EDID information will be present and optimized so mode list creation
and taversal time is minimized.

Add a new module parameter that replaces the normal fbdev initialization
with a new initial mode set.  When the parameter is off, follow the normal
program flow and (if configured) initialize fbdev and a framebuffer
console.  When the parameter is on, do an atomic mode set with everything
off (no crtc's enabled).

This creates the framework for the modesetting and framebuffer code
in the following commits.

v2: Move comment down and format in kerneldoc format (Matt)
    Remove unused include files (Matt)
    Remove unused variables (Bob)
    Major code reorganizations and remove dependencies on fb helper code.
v3: Remove the modeset_info structure.
    Merge the initial modesetting code to create a complete implementation.
    Rename width/height to max_width/max_height in set_display_mode (Matt)
    Fix out of date comment about allocating empty framebuffer (Matt)
v4: Remove the modeset structure abstraction and use atomic state directly.
v5: Remove redundent call to drm_atomic_add_affected_connectors(). (Chi)
    Fix atomic commit locking. (Bob)
v6: Remove the modeset_info structure and pass state around instead. (Matt)
v7: Change the initial split of the first two commits.  This commit now
    sets an initial config with everything off.  The next commit adds
    checking connector status and turns on those with attached displays.
v8: Tweak the docbook info and add to drm.tmpl (Matt)
    Change disable_planes to propogate errors (Matt)
    Remove the expliced setting of disabled planes crtc & src rect (Matt)
    Remove any old framebuffers after disabling the plane (Bob)
    Move the atomic commit and uplevel the backoff code handling (Matt)
v9: Property handle initial framebuffer dereference here by moving it
    up from a later patch.
    Rename file to intel_initial_modeset.c to be less confusing (Matt)
    Change module paramter to a boolean (Matt)
    Use a workqueue instead of async schedule (Matt)
v10: remove double ;; (Matt)
     Change DRM_ERROR to DRM_DEBUG_KSM on deadlock (Matt)

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: James Xiong <james.xiong@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>