iio: common: hid-sensors: Change the precision of scale values
Extension of the scale precision is needed as there are some scale values
that exceed 6 decimal points precision. For example, the scale for
gyroscope from degrees/second to radians/second is 0.017453. If the unit
exponent for gyroscope is -5, it will be
0.017453 -> exp: -5 -> 0.00000017453

However, because the decimal precision is up to 6, the mentioned scale
value will be truncated to 0.000000. To avoid this, changes are made to
extend the precision to 9 decimal points precision.

Signed-off-by: Ooi, Joyce <joyce.ooi@intel.com>