ASoC: Intel: Skylake: Support for DSP exception record dump
In the cases where the DSP encounters an exception
during its execution, the record is stored in the FW
registers window, aligned towards the end. This data
is read by the driver and is passed to the user space
using the linux coredump framework.
The record contains data on a per core basis is dumped to
the userspace in two scenarios -
1. when DSP sends an EXCEPTION_CAUGHT IPC
2. during any IPC timeout

Change-Id: I3fbb50fc705fbb593f03b6a556fab26e8eed8c30
Signed-off-by: Mousumi Jana <mousumix.jana@intel.com>