drm/i915: Ensure splash image requests work in any display order
When a splash screen image is shared by multiple displays, it is
loaded only once for the first splash entry, an gem obj is then
created and shared with other entries.
In this case, eDP is the 2nd entry in the boot parameter, but gets
initialized first. Because eDP->fw is NULL, it failed to create
fb for the eDP.

The fix is to remove the unnecessary check in create_splash_fb.

Signed-off-by: Xiong, James <james.xiong@intel.com>
Reviewed-by: Bob Paauwe <bob.j.paauwe@intel.com>
[mattrope: Cleaned up commit message/headline]
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>