drm/i915: Do not delay HDMI unplug detection
To work around  the issue that live status register takes
some time to settle we implemented a de-bounce filter in
the hdmi detection code that can take up to 80ms to detect
a successful plug-in detection.

The way this was implement meant that upon an unplug event
the driver would always take 80ms to detect it. This long delay
is causing us to fail a HDMI conformance test that requires that
the port be taken down within 100ms of an unplug.

The solution is to report the unplug event straight way when we
know that the connector was plugged in, thus giving the rest of the
driver the maximum amount we can to meet the 100ms conformance
requirement.

V2: Addressed review comments by Raf
	- Replace the commit message with the suggested one
	- Change the comment
	- Fix the pre-increment in while loop

Signed-off-by: Shashank Sharma <shashank.sharma@intel.com>
Reviewed-by: Rafael Barbalho <rafael.barbalho@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>